refactor(api): move validation error serializer to src
