+++ api/src/evaluation/application/pre-handlers/assessment-authorization.js
@@ -1,5 +1,5 @@
-import * as validationErrorSerializer from '../../../../lib/infrastructure/serializers/jsonapi/validation-error-serializer.js';
 import * as assessmentRepository from '../../../shared/infrastructure/repositories/assessment-repository.js';
+import * as validationErrorSerializer from '../../../shared/infrastructure/serializers/jsonapi/validation-error-serializer.js';
 import * as requestResponseUtils from '../../../shared/infrastructure/utils/request-response-utils.js';
 
 const verify = function (
