+++ api/tests/shared/unit/infrastructure/serializers/jsonapi/validation-error-serializer_test.js
@@ -1,6 +1,6 @@
-import { Bookshelf } from '../../../../../lib/infrastructure/bookshelf.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/validation-error-serializer.js';
-import { expect } from '../../../../test-helper.js';
+import { Bookshelf } from '../../../../../../lib/infrastructure/bookshelf.js';
+import * as serializer from '../../../../../../src/shared/infrastructure/serializers/jsonapi/validation-error-serializer.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | validation-error-serializer', function () {
   describe('#serialize', function () {
